+++ api/models/Feedback.py
@@ -2,8 +2,8 @@ from django.db import models
 
 
 class Feedback(models.Model):
-    bus_id = models.IntegerField()
-    bus_line = models.CharField()
+    bus_id = models.IntegerField(null=True)
+    bus_line = models.CharField(null=True, max_length=3)
     passenger_density = models.IntegerField(
         verbose_name="Passenger Density",
         choices=[
