+++ api/views/__init__.py
@@ -10,3 +10,4 @@ from .search_service_views import (AvailableLineView,
 from .obstacle_marker import ObstacleMarkerView
 from .buses_location_view import MockupBusesLocationListView, PredictedBusDataView
 from .passenger_density_view import DensityView
+from .save_feedback_view import SaveFeedbackView
